backend/requirements.txt
uvicorn
fastapi
stripe
pydantic-settings
python-dotenv
jinja2
pytest
pytest-mock
sqlalchemy
httpx
python-jose
psycopg2-binary
passlib
bcrypt==4.0.1 
python-multipart
emails
aiofiles
cloudinary